@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 
 import '../../domain/models/theme_type.dart';
 import '../values/palette.dart';
@@ -29,17 +28,17 @@ class DynamicTheme extends StatefulWidget {
   }
 
   // ---------------------------------------------------------------------------
-  static _DynamicThemeState? instanceOf(BuildContext context) {
+  static _DynamicThemeState instanceOf(BuildContext context) {
     final _DynamicThemeInherited? inherited =
         context.dependOnInheritedWidgetOfExactType<_DynamicThemeInherited>();
-    return inherited?.data;
+    return inherited!.data;
   }
 
   // ---------------------------------------------------------------------------
   static Palette paletteOf(BuildContext context) {
     final _DynamicThemeInherited? inherited =
         context.dependOnInheritedWidgetOfExactType<_DynamicThemeInherited>();
-    return inherited?.data.palette ?? Palette.day();
+    return inherited?.data.palette ?? Palette.light();
   }
 }
 
@@ -59,6 +58,10 @@ class _DynamicThemeState extends State<DynamicTheme> {
 
   ThemeTypes get themeType => _themeType;
 
+  bool get isLightTheme => _themeType == ThemeTypes.light;
+
+  bool get isDarkTheme => _themeType == ThemeTypes.dark;
+
   // ---------------------------------------------------------------------------
   @override
   void initState() {
@@ -77,12 +80,13 @@ class _DynamicThemeState extends State<DynamicTheme> {
     if (widget.initialThemeKey == null && !isThemeInitialized) {
       final Brightness systemBrightness =
           MediaQuery.of(context).platformBrightness;
+
       if (systemBrightness == Brightness.dark) {
-        _themeType = ThemeTypes.night;
+        _themeType = ThemeTypes.dark;
         _palette = ThemeBuilder.getPalette(_themeType);
         _theme = ThemeBuilder.getTheme(_themeType, _palette);
       } else {
-        _themeType = ThemeTypes.day;
+        _themeType = ThemeTypes.light;
         _palette = ThemeBuilder.getPalette(_themeType);
         _theme = ThemeBuilder.getTheme(_themeType, _palette);
       }
@@ -104,34 +108,35 @@ class _DynamicThemeState extends State<DynamicTheme> {
   @override
   Widget build(BuildContext context) {
     return _DynamicThemeInherited(
-        data: this,
-        child: AnnotatedRegion<SystemUiOverlayStyle>(
-          value: SystemUiOverlayStyle(
-            /// The brightness of the top status bar icons.
-            /// Only honored in Android version M and greater.
-            statusBarIconBrightness: _themeType != ThemeTypes.night
-                ? Brightness.dark
-                : Brightness.light,
-
-            /// The brightness of top status bar.
-            /// Only honored in iOS.
-            statusBarBrightness: _themeType != ThemeTypes.night
-                ? Brightness.light
-                : Brightness.dark,
-
-            /// The color of the system bottom navigation bar.
-            /// Only honored in Android versions O and greater.
-            systemNavigationBarColor: _palette.black1,
-
-            /// The brightness of the system navigation bar icons.
-            /// Only honored in Android versions O and greater.
-            systemNavigationBarIconBrightness:
-                _theme.brightness == Brightness.dark
-                    ? Brightness.dark
-                    : Brightness.light,
-          ),
-          child: widget.child,
-        ));
+      data: this,
+      // child: AnnotatedRegion<SystemUiOverlayStyle>(
+      // value: SystemUiOverlayStyle(
+      //   /// The brightness of the top status bar icons.
+      //   /// Only honored in Android version M and greater.
+      //   statusBarIconBrightness: _themeType != ThemeTypes.dark
+      //       ? Brightness.dark
+      //       : Brightness.light,
+      //
+      //   /// The brightness of top status bar.
+      //   /// Only honored in iOS.
+      //   statusBarBrightness: _themeType != ThemeTypes.dark
+      //       ? Brightness.light
+      //       : Brightness.dark,
+      //
+      //   /// The color of the system bottom navigation bar.
+      //   /// Only honored in Android versions O and greater.
+      //   systemNavigationBarColor: _palette.background1,
+      //
+      //   /// The brightness of the system navigation bar icons.
+      //   /// Only honored in Android versions O and greater.
+      //   systemNavigationBarIconBrightness:
+      //       _theme.brightness == Brightness.dark
+      //           ? Brightness.dark
+      //           : Brightness.light,
+      // ),
+      child: widget.child,
+      // )
+    );
   }
 }
 
